PythonStudy/PythonApp/PythonApp/requirements.txt
certifi==2019.3.9
chardet==3.0.4
Django==1.11.20
idna==2.8
mysql-connector==2.2.9
pip==19.1
pytz==2019.1
requests==2.21.0
setuptools==41.0.1
urllib3==1.24.2
wheel==0.33.1
